Services/PongServer/requirements.txt
pyjwt
websockets[12.0]
python-decouple

